pyproject.toml
[project]
name = "pdf-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "img2pdf>=0.6.3",
    "pillow>=12.0.0",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
]

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]

[dependency-groups]
dev = [
    "pyrefly>=0.43.1",
    "ruff>=0.14.7",
]

[tool]

# --- Custom Config ---
[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
